+++ backend/src/main/resources/application.properties
@@ -4,8 +4,8 @@ server.port=8080
 #security
 application.security.jwt.secretkey=404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970
 application.security.jwt.token-prefix=Bearer 
-# a day in milliseconds
-application.security.jwt.expiration=86400000
+# a week in milliseconds
+application.security.jwt.expiration=604800000
 
 spring.servlet.multipart.max-file-size=100MB
 spring.servlet.multipart.max-request-size=100MB
