+++ frontend/src/pages/UserProfilePage.tsx
@@ -1,28 +1,38 @@
 import React, {useEffect, useState} from "react";
 import VideoCard from "../components/VideoCard";
 import Modal from '../components/Modal';
-import {CommentDTO} from "../types/videoInterfaces.tsx";
-import {getCommentsByAuthor} from "../services/videoService.ts";
-import {useUser} from "../contexts/UserContext.tsx"; // Asegúrate de ajustar la ruta de importación según tu estructura de proyecto
+import {CommentDTO, VideosDTO} from "../types/videoInterfaces.tsx";
+import {getVideosByAuthor, getCommentsByAuthor} from "../services/videoService.ts";
+import {useUser} from "../contexts/UserContext.tsx";
+import { getImagesForVideos } from "../utils/functions.ts";
 
 const UserProfilePage: React.FC = () => {
   const [isUploadModalOpen, setUploadModalOpen] = useState(false);
   const [videoTitle, setVideoTitle] = useState('');
   const [videoFile, setVideoFile] = useState<File | null>(null);
   const [comments, setComments] = useState<CommentDTO[]>([]);
+  const [videos, setVideos] = useState<VideosDTO[]>([]);
   const {currentUser} = useUser();
+  const [imageUrls, setImageUrls] = useState<{ [key: number]: string }>({});
 
   useEffect(() => {
-    const fetchComments = async () => {
+    const fetchData = async () => {
       try {
         const commentsData = await getCommentsByAuthor();
         setComments(commentsData);
+
+        const videosData = await getVideosByAuthor();
+        setVideos(videosData);
+
+        const imagesUrlsMap = await getImagesForVideos(videosData);
+        setImageUrls(imagesUrlsMap);
       } catch (error) {
-        console.error('Error fetching comments:', error);
+        console.error('Error fetching data:', error);
       }
     };
 
-    fetchComments().then(r => console.log('Comments fetched:', r));
+    fetchData().then(r => console.log('Data fetched:', r));
+
   }, []);
 
   const handleUploadVideo = () => {
@@ -44,11 +54,11 @@ const UserProfilePage: React.FC = () => {
   };
 
   // Ejemplo de datos de videos del usuario
-  const userVideos = [
+  /*const userVideos = [
     { id: 1, title: 'Mi primer video', owner: 'Jane Doe', imagePath: '/assets/video1-thumbnail.png' },
     { id: 2, title: 'Tutorial de React', owner: 'Jane Doe', imagePath: '/assets/video2-thumbnail.png' },
     // Añade más videos según sea necesario
-  ];
+  ];*/
 
   return (
     <div className="flex flex-col items-center bg-gray-50 min-h-screen py-10">
@@ -78,18 +88,25 @@ const UserProfilePage: React.FC = () => {
       {/* Sección de videos */}
       <div className="bg-white w-full max-w-4xl p-6 rounded-lg shadow-lg mb-6">
         <h2 className="text-xl font-bold mb-4 text-gray-800">My Videos</h2>
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-          {userVideos.map(video => (
-            <VideoCard key={video.id} {...video} />
-          ))}
-        </div>
+          {videos.length === 0 ? (
+            <div className="flex justify-center items-center h-32">
+              <p className="text-gray-600 text-sm text-center">No videos yet</p>
+            </div>
+          ) : (
+            <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+              {videos.map(video => (
+                <VideoCard key={video.id} {...video} imagePath={imageUrls[video.id]}/>
+              ))}
+            </div>
+          )}
       </div>
 
       {/* Sección de comentarios */}
       <div className="bg-white w-full max-w-4xl p-6 rounded-lg shadow-lg">
         <h2 className="text-xl font-bold mb-4 text-gray-800">My Comments</h2>
         {/* Tarjeta de comentario */}
-        {comments.map((comment, index) => (
+        {comments.length === 0 && <p className="text-gray-600 text-sm text-center">No comments yet</p>}
+          {comments.map((comment, index) => (
             <div key={index} className="bg-gray-100 rounded-lg p-4 mb-4 shadow">
               <p className="text-gray-600 text-sm text-left">{comment.comment_text}</p>
             </div>
